health-monitoring-system/requirements.txt
flask
pandas
numpy
scikit-learn
pycryptodome